moving the menu function to server.py - still not working
