+++ client.py
@@ -2,97 +2,7 @@
 
 import socket
 import sys
-def menu():
-    d = {}
-    with open(r"C:\Users\SREEJITH\Desktop\IITGN\DB\Database.txt") as f:
-        for line in f:
-           (key, val) = line.split()
-           d[key] = val
-    def listfile():
-        pass
-    def uploadfile():
-        pass
-    def downloadfile():
-        pass
-    def deletefile():
-        pass
-    def sharefile():
-        pass
-    def showlog():
-        pass
-    def signup():
-        s=1
-        while s:
-            username  = input("Enter a username:")
-            s=0
-            for i in d:
-                    if i==username:
-                        print("Username already exists! Choose a new one")
-                        s=1
-        while True:
-            password  = input("Enter a password:")
-            password2 = input("Confirm password:")
-            if password == password2:
-                file = open(r"C:\Users\SREEJITH\Desktop\IITGN\DB\Database.txt", "a+")
-                #file = open(r"file.txt", "a+")
-                file.write(username+" "+password+"\n")
-                file.close()
-                break
-            print("Passwords don't match!")
-    def login():
-        s=0
-        while True:
-            username = input("Login:")
-            password = input("Password:")
-            file = open(r"C:\Users\SREEJITH\Desktop\IITGN\DB\Database.txt", "r")
-            #file = open(r"file.txt", "r")
-            for i in d:
-                if i==login1 and d[i]==login2:
-                    s=1
-                    break
-            if s==1:
-                break    
-            else:
-                print("Incorrect Password!")
-                
-
-    print("Welcome...")
-    while 1:
-        welcome = int(input("1.Sign up\n2.Sign in\nEnter your choice(1-2): "))
-        if welcome == 1:
-            signup()
-            break
-        if welcome == 2:
-            login()
-            break
-        if welcome>2 or welcome<1:
-            print("Incorrect Choice")
-    while 1:
-        print("\n1.List files:")
-        print("\n2.Upload files:")
-        print("\n3.Download files:")
-        print("\n4.Delete files:")
-        print("\n5.Share files:")
-        print("\n6.Show log:")
-        print("\n7.Sign out:")
-        choice=int(input("Enter choice(1-7): "))
-        if(choice==1):
-            listfile()
-        if(choice==2):
-            uploadfile()
-        if(choice==3):
-            downloadfile()
-        if(choice==4):
-            deletefile()
-        if(choice==5):
-            sharefile()
-        if(choice==6):
-            showlog()
-        if(choice==7):
-            print("Thank you")
-            return 'quit'
-        if(choice>7 or choice<1):
-            print("Incorrect choice")
+import getpass
 def clientConnect(hostStr, dataPort, cmdPort):
 
     #create socket object
